requirements.txt
jupyter>=1.0,<2.0
ipykernel>=6.0,<7.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
numpy>=1.26,<3.0
# smolagents>=1.0,<2.0  # Phase 2 — uncomment when core/06 is added
